requirements.txt
pymysql==1.1.0
gunicorn==21.2.0
cryptography==41.0.7
urllib3==2.0.7
PyMySQL[rsa]==1.1.0
Flask==2.3.3
flask-cors==4.0.0
requests==2.32.3
mysql-connector-python==8.4.0
psycopg2-binary==2.9.9
